继承的mode id变为uint64

+++ model/blog_model.go
@@ -8,10 +8,10 @@ const (
 
 type BlogModel struct {
 	BaseModel
-	UserID      uint   `gorm:"cloumn:user_id"` //userid
+	UserID      uint64   `gorm:"cloumn:user_id"` //userid
 	DocID       string `gorm:"cloumn:doc_id"`  //es文档id
-	CoverPlanId int64  `gorm:"cloumn:cover_plan_id"`
-	BlogTypeId  int64  `gorm:"cloumn:blog_type_id"`
+	CoverPlanId uint64  `gorm:"cloumn:cover_plan_id"`
+	BlogTypeId  uint64  `gorm:"cloumn:blog_type_id"`
 	YuqueId     int64  `gorm:"cloumn:yuque_id"` //语雀文档id
 	YuqueSlug   string `gorm:"cloumn:yuque_slug"`
 	YuqueFormat string `gorm:"cloumn:yuque_format"`
